capturesync/requirements.txt
watchdog
Pillow
rich
